Updated users method with cache process

+++ src/Controller/UserController.php
@@ -3,40 +3,81 @@
 namespace App\Controller;
 
 use App\Entity\User;
-use App\Repository\CustomerRepository;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Controller\Annotations\Get;
 use FOS\RestBundle\Controller\Annotations\View;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Contracts\Cache\ItemInterface;
+use Symfony\Contracts\Cache\TagAwareCacheInterface;
 
-
+/**
+ * Permet de gérer les requêtes sur les utilisateurs
+ * Class UserController
+ * @package App\Controller
+ */
 class UserController extends AbstractController
 {
+    private $userRepo;
+    private $requestedUser;
+
+    /**
+     * UserController constructor.
+     * @param UserRepository $repository
+     */
+    public function __construct(UserRepository $repository)
+    {
+        $this->userRepo = $repository;
+
+    }
+
     /**
      * @Get("/users", name="list_users")
-     * @View
-     * @param UserRepository $userRepository
-     * @return User[]
+     * @View(StatusCode = 200)
+     * @param TagAwareCacheInterface $cache
+     * @return mixed
+     * @throws \Psr\Cache\InvalidArgumentException
+     * @Security("is_granted('ROLE_USER') ")
      */
-    public function users(UserRepository $userRepository)
+    public function users(TagAwareCacheInterface $cache)
     {
-        $allUsers = $userRepository->findAll();
-        return $allUsers;
+        return $cache->get('users', function (ItemInterface $item) {
+            $item->expiresAfter(1800);
+            $customer = $this->getUser();
+            $usersOfCustomer = $this->userRepo->findBy(['customer' => $customer]);
+            return $usersOfCustomer;
+        });
+
     }
 
 
     /**
      * @Get("/users/{id}", name="one_user")
-     * @View
+     * @View(StatusCode = 200)
      * @param User $user
      * @return User
+     * @Security("is_granted('ROLE_USER') ")
      */
-    public function getOneUser(User $user)
+    public function getOneUser(User $user, TagAwareCacheInterface $cache)
     {
-        return $user;
+        $this->setRequestedUser($user);
+
+        return $cache->get('users'.$this->requestedUser->getId(), function (ItemInterface $item) {
+            $item->expiresAfter(1800);
+
+            if ($this->getUser()->getId() == $this->getRequestedUser()->getCustomer()->getId()) {
+                return $this->getRequestedUser();
+            } else {
+                return new Response('Cet utilisateur ne vous appartient pas',403);
+            }
+        });
     }
 
     /**
@@ -45,42 +86,90 @@ class UserController extends AbstractController
      * @ParamConverter("user", converter="fos_rest.request_body")
      * @param User $user
      * @param EntityManagerInterface $entityManager
-     * @param CustomerRepository $customerRepository
-     * @return User
+     *
+     * @return User|Response
+     * @Security("is_granted('ROLE_USER') ")
      */
-    public function createUser(User $user, EntityManagerInterface $entityManager, CustomerRepository $customerRepository)
+    public function createUser(Request $request, User $user,
+                               EntityManagerInterface $entityManager, TagAwareCacheInterface $cache,
+                               ValidatorInterface $validator)
     {
         $user->setCustomer($this->getUser());
 
         $entityManager->persist($user);
         $entityManager->flush();
-
+        $cache->delete('users');
         return $user;
     }
 
 
     /**
      * @Rest\Put("/users/{id}", name="update_user")
+     * @View(StatusCode = 200)
+     * @ParamConverter("updatedUser", converter="fos_rest.request_body")
      * @param User $user
+     * @param User $updatedUser
      * @param EntityManagerInterface $entityManager
-     * @return User
+     * @param TagAwareCacheInterface $cache
+     * @return User|Response
+     * @Security("is_granted('ROLE_USER') ")
      */
-    public function updateUser(User $user, EntityManagerInterface $entityManager)
+    public function updateUser(User $user, User $updatedUser, EntityManagerInterface $entityManager, TagAwareCacheInterface $cache)
     {
+        $requestedUser = $user;
+        $cache->delete('users');
+        if ($this->getUser()->getId() == $requestedUser->getCustomer()->getId()) {
 
-        $entityManager->flush();
+            $user->setCustomer($this->getUser());
+            $user->setFirstname($updatedUser->getFirstname());
+            $user->setLastname($updatedUser->getLastname());
+            $user->setEmail($updatedUser->getEmail());
 
-        return $user;
+            $entityManager->flush();
+            return $user;
+
+        } else {
+            return new Response('Cet utilisateur ne vous appartient pas');
+        }
     }
 
     /**
      * @Rest\Delete("/users/{id}", name="delete_user")
+     * @View(StatusCode = 200)
      * @param User $user
      * @param EntityManagerInterface $entityManager
+     * @Security("is_granted('ROLE_USER') ")
      */
-    public function deleteUser(User $user, EntityManagerInterface $entityManager)
+    public function deleteUser(User $user, EntityManagerInterface $entityManager, TagAwareCacheInterface $cache)
     {
-        $entityManager->remove($user);
+        $cache->delete('users'.$this->requestedUser->getId());
+        $requestedUser = $user;
+        if ($this->getUser()->getId() == $requestedUser->getCustomer()->getId()) {
+            $entityManager->remove($user);
+            $entityManager->flush();
+
+        } else {
+            return new Response('Cet utilisateur ne vous appartient pas');
+        }
+        return new JsonResponse(['message' => 'Utilisateur supprimé']);
+}
+
+    /**
+     * @return mixed
+     */
+    public function getRequestedUser()
+    {
+        return $this->requestedUser;
+    }
+
+    /**
+     * @param mixed $requestedUser
+     */
+    public function setRequestedUser($requestedUser)
+    {
+        $this->requestedUser = $requestedUser;
+
+        return $this;
     }
 
 }
